# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bookmark Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tag {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
  sortOrder: number;
}

export interface Bookmark {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  previewImage?: string | null;
  /** @nullable */
  collectionId?: number | null;
  pinned: boolean;
  sortOrder?: number;
  createdAt: string;
  updatedAt?: string;
  tags?: Tag[];
}

export interface BookmarkInput {
  url: string;
  title: string;
  description?: string;
  favicon?: string;
  previewImage?: string;
  /** @nullable */
  collectionId?: number | null;
  pinned?: boolean;
  tagIds?: number[];
}

export interface BookmarkUpdate {
  url?: string;
  title?: string;
  description?: string;
  favicon?: string;
  previewImage?: string;
  /** @nullable */
  collectionId?: number | null;
  pinned?: boolean;
  sortOrder?: number;
  tagIds?: number[];
}

export interface BookmarkStats {
  total: number;
  pinned: number;
  collections: number;
  tags: number;
  recentCount?: number;
}

export interface Collection {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  parentId?: number | null;
  createdAt: string;
  bookmarkCount?: number;
}

export interface CollectionInput {
  name: string;
  description?: string;
  color?: string;
  icon?: string;
  /** @nullable */
  parentId?: number | null;
}

export interface CollectionUpdate {
  name?: string;
  description?: string;
  color?: string;
  icon?: string;
  /** @nullable */
  parentId?: number | null;
}

export interface TagInput {
  name: string;
  color?: string;
}

export interface TagUpdate {
  name?: string;
  /** @nullable */
  color?: string | null;
  sortOrder?: number;
}

export interface ImportInput {
  /** Raw HTML content from Firefox/Chrome bookmark export */
  html: string;
  /**
     * Optional collection to import into
     * @nullable
     */
  collectionId: number | null;
}

export interface ImportResult {
  imported: number;
  skipped: number;
  errors: number;
  bookmarks?: Bookmark[];
}

export type ListBookmarksParams = {
/**
 * @nullable
 */
collectionId?: number | null;
/**
 * @nullable
 */
tagId?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
pinned?: boolean | null;
};