# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bookmark Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Tag } from './tag';

export interface Bookmark {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  previewImage?: string | null;
  /** @nullable */
  collectionId?: number | null;
  pinned: boolean;
  sortOrder?: number;
  createdAt: string;
  updatedAt?: string;
  tags?: Tag[];
}