# lib/api-zod/src/generated/types/bookmarkInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bookmark Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BookmarkInput {
  url: string;
  title: string;
  description?: string;
  favicon?: string;
  previewImage?: string;
  /** @nullable */
  collectionId?: number | null;
  pinned?: boolean;
  tagIds?: number[];
}