# lib/api-zod/src/generated/types/tag.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bookmark Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Tag {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
  sortOrder: number;
}